lightserv/requirements.txt
acronym==1.6.0
acstools==2.0.10
alabaster==0.7.9
APLpy==1.1.1
appdirs==1.4.0
appnope==0.1.0
asdf==1.1.0.dev728
asn1crypto==0.24.0
astroimtools==0.2.dev177
astropy==2.0.2
astroquery==0.3.3
asv==0.1.1
Babel==2.3.4
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
bcrypt==3.1.4
beautifulsoup4==4.5.1
bkcharts==0.2
bleach==1.5.0
bokeh==0.12.7
calcos==3.3.4
certifi==2016.2.28
cffi==1.11.5
click==6.7
cloudpickle==0.4.0
configparser==3.5.0
corner==2.0.2.dev0
CosmoloPy==0.1.104
costools==1.2.1.dev0
crds==7.2.4
cryptography==2.3.1
Cube-Tools==0.0.1
cycler==0.10.0
Cython==0.25.1
d2to1==0.2.12
dask==0.15.2
decorator==4.0.9
dill==0.2.6
docopt==0.6.2
docutils==0.12
drizzlepac==2.2.6
emcee==2.2.1
entrypoints==0.2.3
enum34==1.1.6
ephem==3.7.6.0
filelock==2.0.7
fitsblender==0.3.2
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-Test==0.1.6
Flask-Testing==0.7.1
Flask-WTF==0.14.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.1.1
ginga==2.5.20161005204600
glue-core==0.12.0
glue-vispy-viewers==0.9
glueviz==0.10.0
grizli==0.2.1
h5py==2.6.0
htc-utils==0.1
html5lib==0.999999999
idna==2.7
imagesize==0.7.1
imexam==0.8.0
ipaddress==1.0.22
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==7.4.2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.2.0
jupyter-core==4.2.0
keyring==10.0.2
lacosmic==0.1.1
locket==0.2.0
lxml==3.8.0
MarkupSafe==0.23
matplotlib==1.5.3
mechanize==0.3.2
mistune==0.7.4
mock==2.0.0
nbconvert==5.2.1
nbformat==4.4.0
networkx==1.11
nictools==1.1.3.dev0
nltk==3.4
nose==1.3.7
notebook==5.0.0
numpy==1.11.2
numpydoc==0.6.0
observer==0.1.0
pandas==0.19.0
pandocfilters==1.4.2
paramiko==2.4.1
Parsley==1.3
partd==0.3.8
path.py==0.0.0
pathlib2==2.1.0
pbr==1.10.0
PeakUtils==1.0.3
pexpect==4.0.1
photutils==0.2.2
pickleshare==0.7.4
Pillow==3.4.2
pkgconfig==1.1.0
plotly==2.0.11
prompt-toolkit==1.0.8
ptyprocess==0.5.1
purge-path==1.0.0
py==1.4.31
py-applescript==1.0.0
py-expression-eval==0.3.3
pyasn1==0.4.4
pycparser==2.18
pyds9==1.9.dev127
pyFFTW==0.9.2
pyfits==3.5
Pygments==2.1.3
PyNaCl==1.2.1
pyobjc==3.2.1
pyobjc-core==3.2.1
pyobjc-framework-Accounts==3.2.1
pyobjc-framework-AddressBook==3.2.1
pyobjc-framework-AppleScriptKit==3.2.1
pyobjc-framework-AppleScriptObjC==3.2.1
pyobjc-framework-ApplicationServices==3.2.1
pyobjc-framework-Automator==3.2.1
pyobjc-framework-AVFoundation==3.2.1
pyobjc-framework-AVKit==3.2.1
pyobjc-framework-CalendarStore==3.2.1
pyobjc-framework-CFNetwork==3.2.1
pyobjc-framework-CloudKit==3.2.1
pyobjc-framework-Cocoa==3.2.1
pyobjc-framework-Collaboration==3.2.1
pyobjc-framework-Contacts==3.2.1
pyobjc-framework-ContactsUI==3.2.1
pyobjc-framework-CoreBluetooth==3.2.1
pyobjc-framework-CoreData==3.2.1
pyobjc-framework-CoreLocation==3.2.1
pyobjc-framework-CoreText==3.2.1
pyobjc-framework-CoreWLAN==3.2.1
pyobjc-framework-CryptoTokenKit==3.2.1
pyobjc-framework-DictionaryServices==3.2.1
pyobjc-framework-DiskArbitration==3.2.1
pyobjc-framework-EventKit==3.2.1
pyobjc-framework-ExceptionHandling==3.2.1
pyobjc-framework-FinderSync==3.2.1
pyobjc-framework-FSEvents==3.2.1
pyobjc-framework-GameCenter==3.2.1
pyobjc-framework-GameController==3.2.1
pyobjc-framework-ImageCaptureCore==3.2.1
pyobjc-framework-IMServicePlugIn==3.2.1
pyobjc-framework-InputMethodKit==3.2.1
pyobjc-framework-InstallerPlugins==3.2.1
pyobjc-framework-InstantMessage==3.2.1
pyobjc-framework-IOSurface==3.2.1
pyobjc-framework-LatentSemanticMapping==3.2.1
pyobjc-framework-LaunchServices==3.2.1
pyobjc-framework-LocalAuthentication==3.2.1
pyobjc-framework-MapKit==3.2.1
pyobjc-framework-MediaAccessibility==3.2.1
pyobjc-framework-MediaLibrary==3.2.1
pyobjc-framework-ModelIO==3.2.1
pyobjc-framework-MultipeerConnectivity==3.2.1
pyobjc-framework-NetFS==3.2.1
pyobjc-framework-NetworkExtension==3.2.1
pyobjc-framework-NotificationCenter==3.2.1
pyobjc-framework-OpenDirectory==3.2.1
pyobjc-framework-Photos==3.2.1
pyobjc-framework-PhotosUI==3.2.1
pyobjc-framework-PreferencePanes==3.2.1
pyobjc-framework-PubSub==3.2.1
pyobjc-framework-QTKit==3.2.1
pyobjc-framework-Quartz==3.2.1
pyobjc-framework-SafariServices==3.2.1
pyobjc-framework-SceneKit==3.2.1
pyobjc-framework-ScreenSaver==3.2.1
pyobjc-framework-ScriptingBridge==3.2.1
pyobjc-framework-SearchKit==3.2.1
pyobjc-framework-ServiceManagement==3.2.1
pyobjc-framework-Social==3.2.1
pyobjc-framework-SpriteKit==3.2.1
pyobjc-framework-StoreKit==3.2.1
pyobjc-framework-SyncServices==3.2.1
pyobjc-framework-SystemConfiguration==3.2.1
pyobjc-framework-WebKit==3.2.1
PyOpenGL==3.1.1a1
pyparsing==2.1.4
pyqtgraph==0.9.11.dev0
pyregion==1.1.2
pysynphot==0.9.12
pytest==3.0.3
python-dateutil==2.5.3
python2-secrets==1.0.5
pytools==2016.1
pytz==2016.7
pywcs==1.12
PyYAML==3.12
pyzmq==16.0.0
qtconsole==4.2.1
QtPy==1.3.1
reftools==1.7.4
requests==2.11.1
scikit-image==0.12.3
scikit-learn==0.18
scipy==0.18.1
seaborn==0.9.0
selenium==3.14.1
sewpy==1.0.dev0
Shapely==1.5.16
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
specutils==0.2.2
specview==0.2.dev367
specviz==0.4.4
spherical-geometry==1.2.4
Sphinx==1.4.8
sphinx-automodapi==0.8
sphinx-bootstrap-theme==0.4.11
SQLAlchemy==1.2.0b2
stginga==0.1.2.dev197
stistools==1.1
stregion==1.1.4
stsci.convolve==2.2.2
stsci.distutils==0.3.8.dev0
stsci.image==2.3.0
stsci.imagemanip==1.1.2.dev0
stsci.imagestats==1.4.1.dev0
stsci.ndimage==0.10.1.dev0
stsci.numdisplay==1.6.1.dev0
stsci.skypac==0.9.dev0
stsci.sphere==0.2
stsci.sphinxext==1.2.2.dev0
stsci.stimage==0.2.1.dev0
stsci.tools==3.4.13
stwcs==1.2.4
terminado==0.6
testpath==0.3
toolz==0.8.2
tornado==4.4.2
tqdm==4.31.1
traitlets==4.3.1
urllib3==1.24
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.11.15
wfc3tools==1.3.3
wfpc2tools==1.0.3.dev0
widgetsnbextension==3.4.2
WTForms==2.2.1
xlrd==1.0.0
